Splitted security coverage component
