+++ opencti-platform/opencti-graphql/src/modules/playbook/playbook-components.ts
@@ -13,7 +13,6 @@ distributed under the License is distributed on an "AS IS" BASIS,
 WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
 */
 import * as R from 'ramda';
-import { v4 as uuidv4 } from 'uuid';
 import type { JSONSchemaType } from 'ajv';
 import * as jsonpatch from 'fast-json-patch';
 import { type BasicStoreEntityPlaybook, ENTITY_TYPE_PLAYBOOK, type PlaybookComponent } from './playbook-types';
@@ -35,7 +34,6 @@ import {
   ENTITY_TYPE_THREAT_ACTOR,
   INPUT_AUTHORIZED_MEMBERS,
   INPUT_GRANTED_REFS,
-  INPUT_LABELS,
   OPENCTI_ADMIN_UUID,
 } from '../../schema/general';
 import type { BasicStoreRelation, StoreCommon, StoreRelation } from '../../types/store';
@@ -53,14 +51,13 @@ import {
   ENTITY_TYPE_TOOL,
   isStixDomainObjectContainer,
 } from '../../schema/stixDomainObject';
-import type { CyberObjectExtension, StixBundle, StixCoreObject, StixCyberObject, StixDomainObject, StixObject, StixOpenctiExtension } from '../../types/stix-2-1-common';
+import type { CyberObjectExtension, StixBundle, StixCoreObject, StixCyberObject, StixObject, StixOpenctiExtension } from '../../types/stix-2-1-common';
 import { STIX_EXT_OCTI, STIX_EXT_OCTI_SCO } from '../../types/stix-2-1-extensions';
 import { connectorsForPlaybook } from '../../database/repository';
 import { internalFindByIds, fullEntitiesList, fullRelationsList, storeLoadById } from '../../database/middleware-loader';
 import { type BasicStoreEntityOrganization, ENTITY_TYPE_IDENTITY_ORGANIZATION } from '../organization/organization-types';
 import { getEntityFromCache } from '../../database/cache';
 import { logApp } from '../../config/conf';
-import { extractStixRepresentative } from '../../database/stix-representative';
 import { isEmptyField, isNotEmptyField, READ_RELATIONSHIPS_INDICES, READ_RELATIONSHIPS_INDICES_WITHOUT_INFERRED } from '../../database/utils';
 import { stixLoadByIds } from '../../database/middleware';
 import { usableNotifiers } from '../notifier/notifier-domain';
@@ -89,10 +86,10 @@ import { PLAYBOOK_SEND_EMAIL_TEMPLATE_COMPONENT } from './components/send-email-
 import { applyOperationFieldPatch, convertMembersToUsers, extractBundleBaseElement } from './playbook-utils';
 import { PLAYBOOK_DATA_STREAM_PIR } from './components/data-stream-pir-component';
 import { convertStoreToStix_2_1 } from '../../database/stix-2-1-converter';
-import { ENTITY_TYPE_SECURITY_COVERAGE, INPUT_COVERED, type StixSecurityCoverage, type StoreEntitySecurityCoverage } from '../securityCoverage/securityCoverage-types';
 import { pushAll } from '../../utils/arrayUtil';
 import { PLAYBOOK_CONTAINER_WRAPPER_COMPONENT } from './components/container-wrapper-component';
 import { PLAYBOOK_MANIPULATE_KNOWLEDGE_COMPONENT } from './components/manipulate-knowledge-component';
+import { PLAYBOOK_SECURITY_COVERAGE_COMPONENT } from './components/security-coverage-component';
 
 // region built in playbook components
 interface LoggerConfiguration {
@@ -478,116 +475,6 @@ export const createTaskFromCaseTemplates = async (
     }
   }
   return tasks;
-};
-
-interface SecurityCoverageConfiguration {
-  all: boolean;
-  auto_enrichment_disable: boolean;
-  periodicity: string;
-  duration: string;
-  type_affinity: string;
-  platforms_affinity: string[];
-}
-const PLAYBOOK_SECURITY_COVERAGE_COMPONENT_SCHEMA: JSONSchemaType<SecurityCoverageConfiguration> = {
-  type: 'object',
-  properties: {
-    all: { type: 'boolean', $ref: 'Create a security coverage for each element of the bundle (on compatible types)', default: false },
-    auto_enrichment_disable: { type: 'boolean', $ref: 'Force manual coverage (prevent enrichment connectors from running)', default: false },
-    periodicity: { type: 'string', $ref: 'Coverage recurrence (every x)', default: 'P1D' },
-    duration: { type: 'string', $ref: 'Duration', default: 'P30D' },
-    type_affinity: {
-      type: 'string',
-      $ref: 'Type affinity',
-      default: 'ENDPOINT',
-    },
-    platforms_affinity: {
-      type: 'array',
-      uniqueItems: true,
-      default: ['windows', 'linux', 'macos'],
-      $ref: 'Platform(s) affinity',
-      items: { type: 'string', oneOf: [] },
-    },
-  },
-  required: ['periodicity', 'duration', 'type_affinity', 'platforms_affinity'],
-};
-
-const SECURITY_COVERAGE_COMPATIBLE_TYPES = [
-  'report',
-  'grouping',
-  'case-incident',
-  'x-opencti-case-incident',
-  'intrusion-set',
-  'campaign',
-  'incident',
-];
-
-export const PLAYBOOK_SECURITY_COVERAGE_COMPONENT: PlaybookComponent<SecurityCoverageConfiguration> = {
-  id: 'PLAYBOOK_SECURITY_COVERAGE_COMPONENT',
-  name: 'Security coverage',
-  description: 'Create a security coverage for the given entity(ies) (when type is compatible)',
-  icon: 'security-coverage',
-  is_entry_point: false,
-  is_internal: true,
-  ports: [{ id: 'out', type: 'out' }],
-  configuration_schema: PLAYBOOK_SECURITY_COVERAGE_COMPONENT_SCHEMA,
-  schema: async () => PLAYBOOK_SECURITY_COVERAGE_COMPONENT_SCHEMA,
-  executor: async ({ dataInstanceId, playbookNode, bundle }) => {
-    const { all, auto_enrichment_disable, periodicity, duration, type_affinity, platforms_affinity } = playbookNode.configuration;
-    const baseData = extractBundleBaseElement(dataInstanceId, bundle) as StixDomainObject;
-    if (SECURITY_COVERAGE_COMPATIBLE_TYPES.includes(baseData.type)) {
-      const name = extractStixRepresentative(baseData);
-      const securityCoverageData: Record<string, unknown> = {
-        name,
-        created: now(),
-        auto_enrichment_disable: auto_enrichment_disable,
-        periodicity: periodicity,
-        duration: duration,
-        type_affinity: type_affinity,
-        platforms_affinity: platforms_affinity,
-        [INPUT_COVERED]: { standard_id: baseData.id },
-        [INPUT_LABELS]: (baseData.labels ?? []).map((l) => ({ value: l })),
-      };
-      const standardId = generateStandardId(ENTITY_TYPE_SECURITY_COVERAGE, securityCoverageData);
-      const storeSecurityCoverage = {
-        internal_id: uuidv4(),
-        standard_id: standardId,
-        entity_type: ENTITY_TYPE_SECURITY_COVERAGE,
-        parent_types: getParentTypes(ENTITY_TYPE_SECURITY_COVERAGE),
-        ...securityCoverageData,
-      } as StoreEntitySecurityCoverage;
-      const securityCoverage = convertStoreToStix_2_1(storeSecurityCoverage) as StixSecurityCoverage;
-      bundle.objects.push(securityCoverage);
-    }
-    if (all) {
-      for (let index = 0; index < bundle.objects.length; index += 1) {
-        const element = bundle.objects[index] as StixDomainObject;
-        if (SECURITY_COVERAGE_COMPATIBLE_TYPES.includes(element.type)) {
-          const name = extractStixRepresentative(element);
-          const securityCoverageData: Record<string, unknown> = {
-            name,
-            created: now(),
-            auto_enrichment_disable: auto_enrichment_disable,
-            periodicity: periodicity,
-            duration: duration,
-            type_affinity: type_affinity,
-            [INPUT_COVERED]: { standard_id: element.id },
-            [INPUT_LABELS]: (element.labels ?? []).map((l) => ({ value: l })),
-          };
-          const standardId = generateStandardId(ENTITY_TYPE_SECURITY_COVERAGE, securityCoverageData);
-          const storeContainer = {
-            internal_id: uuidv4(),
-            standard_id: standardId,
-            entity_type: ENTITY_TYPE_SECURITY_COVERAGE,
-            parent_types: getParentTypes(ENTITY_TYPE_SECURITY_COVERAGE),
-            ...securityCoverageData,
-          } as StoreCommon;
-          const securityCoverage = convertStoreToStix_2_1(storeContainer) as StixSecurityCoverage;
-          bundle.objects.push(securityCoverage);
-        }
-      }
-    }
-    return { output_port: 'out', bundle };
-  },
 };
 
 export interface SharingConfiguration {
